Neumorphism - Basically Finished
I finished all of my html and css so far, for the exception of the custom calendar. Thinking of changing up the code to have everything depend on vh units sop that it is able to scale, but the project says only 1280 by 960 so I'm not sure if I'll do that or not.