Added Hover Effects/Cleaned up Code
I added hover effects on the submit button, as well as the xompany icons. I also cleaned up my code by deleting images that were not being used, but now realize that I might need them in the future. I have a folder with the images saved so it doesn't matter.